add test for multiple console.log in one transaction, move helper code
